requirements.txt
google-generativeai
transformers
torch
pydantic
pyyaml
python-dotenv
